Standardizing logging, Microsoft AD error handling
Standardize error formats to be very verbose, might reconsider and do
verbose logs with more concise (but descriptive) errors. Some errors
print URLs and response bodies, which is useful for debugging but
probably only warranted in verbose logging and not in user-facing
errors.

Also addressed an indexing error when user provides a workspace name
that does not exist.

Wow! This commit is bursting at the seams!
I may have done too much...